added a call to writeObjectArray() in the CDMLineInterface main method.  changed writeObjectArray() method signature to accept List<simulationResult> instead of Object[]

+++ OpenBurn/src/model/SimulationResults.java
@@ -15,7 +15,7 @@ public class SimulationResults
 	private float systemMass;
 	private float systemCenterOfGravity;
 
-	public float getTime()
+	public double getTime()
 	{
 		return time;
 	}
